requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
flake8==7.0.0
pytest==7.4.4